Better definition for xml/sax/__init__.pyi.
Also, merge 2.7/xml/sax and 3/xml/sax into 2and3/xml/sax.
