+++ stdlib/2and3/xml/__init__.pyi
