+++ stdlib/2and3/xml/sax/xmlreader.pyi
